Unbinds Extra from specific Atlantis version
This change weakens the dependency to Atlantis to a compile time
only dependency ('compile' -> 'provided'). This means that there
is no need to release a new version of Atlantis Extra for each
release of Atlantis.

As a consequence the third party app must make sure to provide a
compatible Atlantis dependency in the app.